DEMRUM-3551: use ISO8601DateFormatter and be explicit about extension intent
[repaired]

* Renamed formatter to iso8601Formatter to make it clear this is a standard
  non-localized format suitable for consumption by use cases where that is
  advantageous.
* Updated selected log statements where the code was used to add a localized
  form of the date for human consumption.

Milliseconds are now provided in both formats except on iOS 13 and 14, where
milliseconds will only appear in the standardized format. However, that
format is shown in the log output alongside the localized format, so it will
be easy, in the unusual case where we are running on iOS 13 or 14, to still
see the milliseconds by looking at the iso8601 format on the same line.